+++ src/Schema/Directives/AllDirective.php
@@ -57,7 +57,6 @@ GRAPHQL;
                 $root,
                 $args,
                 $context,
-                $resolveInfo,
             )
             ->get();
     }