+++ src/Schema/Directives/FindDirective.php
@@ -42,7 +42,6 @@ GRAPHQL;
                     $root,
                     $args,
                     $context,
-                    $resolveInfo,
                 )
                 ->get();